+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - (25/02/2021)
+
+### Changed
+- portal-42: Dynamically update rows rather than replacing them each sync.
+- The `uniqueIdColumnName` attribute is required for any AirTable driver configuration.
+
 ## [1.0.4] - (18/06/2020)
 
 ### Added
@@ -38,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AirTable Progress integration
 - AirTable Control Exporter integration
 
-[Unreleased]: https://github.com/bristol-su/control/compare/v1.0.4...HEAD
+[Unreleased]: https://github.com/bristol-su/control/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/bristol-su/control/compare/v1.0.4...v2.0.0
 [1.0.4]: https://github.com/bristol-su/control/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/bristol-su/control/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/bristol-su/control/compare/v1.0.1...v1.0.2